added checking roles
